backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Aurélien Dentan <aurelien.dentan@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.65.2"
alembic = "^1.6.5"
uvicorn = "^0.14.0"
gunicorn = "^20.1.0"
SQLAlchemy = "^1.4.19"
dependency-injector = "^4.33.0"
aioredis = "^1.3.1"
pydantic = {version = "^1.8.2", extras = ["dotenv"]}
httpx = "^0.18.2"
websockets = "^9.1"
pydantic-sqlalchemy = "^0.0.9"
asyncpg = "^0.23.0"
SQLAlchemy-Utils = "^0.37.7"
asgi-ratelimit = {version = "^0.5.1", extras = ["full"]}
opentelemetry-instrumentation-fastapi = "^0.22b0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
black = "^21.6b0"
isort = "^5.8.0"
pre-commit = "^2.13.0"
pylint = "^2.8.3"
coloredlogs = "^15.0.1"
colorama = "^0.4.4"
requests = "^2.25.1"
fakeredis = "^1.5.2"
pytest-asyncio = "^0.15.1"
mypy = "^0.910"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
